refactored for better directory structure
