+++ src/calculator/impact.js
