+++ src/calculator/severe-impact.js
